refactor(extension: podman): extract `checkRosettaMacArm` to dedicated file
